Remove duplicate `govuk::apps::efg_training` config
This duplicates the values immediately above it - I think these were
meant to be set with the key `govuk::apps::efg_training_rebuild`
but as we don't need to override these values in production it's
easier to remove them.

This means the hostname for the rebuild will be
`efg-training-rebuild.publishing.service.gov.uk`.