Measure unprocessed whitehall uploads
Count the number of files in the incoming uploads directory so that we
can see what the virus scanner is doing.

There are also `clean` and `infected` directories we could monitor, but
we were unable to get either of these to work. Both are organised into
many nested subdirectories, and the plugin is supposed to recurse into
these.