Update puppet-aptly
We had an issue with the version of the puppet-aptly module
that caused GPG key errors when bringing up an apt machine. We need to
upgrade to at least 0.6.0, which has a dependency of the puppetlabs-apt module of
>= 2.0.0. This also has an updated dependency of puppetlabs-stdlib.

In version 2.0.0 of puppetlabs-apt, the unattended_upgrades class was
broken out into it's own module, so this installs that as well, and
updates the references in our code to use this as a standalone module.