PostgreSQL Backup Time
Backup fails due to the database not being available.
ERROR: [psql: FATAL:  the database system is shutting down, pg_dumpall: could not connect to database "template1": FATAL:  the database system is shutting down]

The dataset on this cluster is large so running the backup later will give the environment sync more time to complete.