Allow `draft-whitehall-frontend.$app_domain` to serve whitehall assets
In production, the GOV.UK router expects anything under /government to
be served by `whitehall-frontend` unless a more specific route exists.
This includes assets such as uploaded PDFs. This is fine in the live
stack for immigrated content as the Whitehall Rails app will serve the
asset through nginx once it’s performed its own authorisation checks.

For migrated content such as detailed guides, content is previewed
using the draft stack. In this case, the router sends all requests for
/government to `draft-whitehall-frontend`. Before this change, this
would be an invalid vhost in production. This change will create the
vhost and simply serve the files from disk. We make a whitehall-backend
machine serve it as it has access to the NFS mount already. This is a
temporary solution until Whitehall’s asset management system is rebuilt
or ported to use asset-manager.

We don’t need to provide any additional authorisation checking for
these requests as `draft-origin` (the user’s only path to this
endpoint) is protected by `authenticating-proxy`.