Add auth for Github OAuth plugin and VPN
This explicitly calls the relevant classes to set the auth secrets
required to first authenticate with the Github Enterprise VPN, then to
authenticate with Github Enterprise itself.

We pass in parameters and call the classes themselves because when
passing in secrets in encrypted hiera we can only specify per
environment, and not by class or node specific hieradate. Adding
parameters to this class allows us to set different credentials to the
default Integration credentials.

To avoid redeclaring the govuk_jenkins::config class we call it from the
main class, adding the required parameters and updating the default to
call the main class.