Disable licensing smokey test in integration.
Licensing does not contain the correct data in integration, so the
tests have been disabled in smokey itself by using the @notintegration
tag. However, the monitoring check does not respect that tag, so the
failing tests were being run by Icincga anyway.

This change introduces an "enabled" flag in the icinga::check_feature
defined type, defaulting to true, and overrides the check_licensing
hash in integration hieradata to set it to false for that one feature.