Remove manage_config condition
This is set to "true" everywhere, so I don't think it's needed. It was
originally added so we could test the changes in Integration only, but
was never removed when it was set everywhere.

This also recently caused a condition where the changes weren't tested
in Vagrant because the default was "false", and when a test was deployed
to Integration there were unexpected results.