Let govuk::app::envvar handle multiline values
We use envdir (from daemontools) to read the files we write out with
govuk::app::envvar, and this tool only reads the first line of the file.
It does however support multiline values - you have to replace newlines
in the value with null characters.  Envdir will then convert those nulls
back into newlines before setting the value in the environment.

We use an inlinetemplate command to normalize the newlines and then
convert all `\n` into `\0`s.