Add notes about viewing another users token
We came across an issue where we automatically added the user which is
used by Jenkins job builder to authenticate and create jobs, but were
not able to view the users token. To view the token of a user you must
login as that user, but since we do authentication with Github OAuth,
this makes it tricky. This was not an issue on older versions of
Jenkins, but was a security feature that was later introduced.

In lieu of an obvious place to add this documentation, I thought I would
help someone in the future by mentioning in this specific class in case
it crops up again.