Add Jenkins token for job builder
Since we have a second instance of Jenkins we need to be able to pass in
a different API token which is used by Jenkins job builder to manage
jobs. We are unable to differentiate between different classes in the
encrypted hieradata, so we need to split out how we call the class and
call it specifically in the govuk_ci::master class.