Rework attachment syncs
When the work was completed to add in the DR asset slave in the
attachment syncs it caused a condition that can have a significant
impact on the user.

If there is any sort of networking issue to the DR node, then the processing of each
attachment can increase from ~10 seconds to over 2 minutes (due to the 2
minute timeout that is set). There should be no dependency on rsyncing the files across
to both asset slaves before moving onto processing the next attachment
on the asset master. The most important aspect of this process is to
scan the file, and move it on disk on the asset master so that the user
is then able to access the file.

This commit looks to separate the processing and syncing processes.
Instead of including the rsync within the loop, put the file that still
needs to be synced into file. We create a list of files into one file
for each "processing-attachments" loop.

A separate process then goes through all the filelist files and attempts to
sync each file onto both slaves (and an S3 bucket if set). The timeout
has been reduced to 20 seconds because I don't think we will often have
a file which should take longer than 20 seconds to copy. When each file
in a filelist has been copied, the filelist is removed and it moves onto
the next one.

There is also a catch-all task that runs every morning to
ensure the asset master and slaves are all in sync.