Add govuk_ci::job define to use exported resources
We are adding a new job define that we can use to create jobs for specific
apps as exported resources. In the future we can move this to the govuk::app
class so it creates a CI job for each application as default.

Add job only to local-links-manager for testing purposes.