Add firewall rules for need-api new Elasticsearch
We want to switch off the old Elasticsearch machine. This means that
we need to move need-api to the Elasticsearch in the API vDC.

Ideally we would do this by moving need-api to the API vDC:

- Deploy need-api to API machines
- Migrate database and repoint loadbalancers (simultaneously)

But that's a lot more work than just adding these firewall rules for now.