Jenkins swarm client
This commit adds the functionality to connect a Jenkins agent to a Jenkins master with the swarm client.

The swarm client will run on Jenkins agents as an upstart service.

The discovery address is hard coded as a temporary measure until we have a fact to resolve.

At the moment if we wanted to use another version of the swarm client.  We would have to reflect the change in two places with class parameters `$swarm_client_dest` and `$swarm_client_url`.

There is one manual step where we have to retrieve the api_token for the user we're authenticating against the Jenkins master with.
To automate this would mean opening up a potential security hole.
See: [jenkins.security.ApiTokenProperty.showTokenToAdmins](https://wiki.jenkins-ci.org/display/JENKINS/Features+controlled+by+system+properties)