Remove unneeded join
The nagios plugin library check_messages method returns a
collection of all errors and, if none are present, returns all the 'OKs'.
The explicit join seems to be forcing all messages to be collected and joined.
This causes the alert to show all checks, including the passing ones, as failing
if *any* of the checks fail.