Add env vars to Whitehall
Whitehall will be moved to the new deployment pipeline and relevant
configuration needs to be held in environment variables.
Adds the necessary environment variables to configure Whitehall
in both frontend and admin modes. The database credentials are split
according to frontend or admin usage as are some other configuration
vars.
There's an integration-only config value for highlighting words to
avoid when using the admin UI.
Likewise integration needs basic auth credentials for LinkChecker.
The number of unicorn worker processes differs between frontend and
backend configurations.