Enable Jenkins user to run sudo tests on CI
The sudoers.d directory is managed by the sudo module, by default
the Jenkins user can't read the directory so Rake tests are failing
on the new CI boxes.

In this commit we override the resource in the master and agent
classes, eventually we can remove the override from the master.