Add prefix URL to Jenkins
Adding this allows us to create a different URL that is the default for
that Jenkins instance, a requirement that has come about by running two
masters in the same environment.

Keeping 'app_domain' as a parameter will allow us to override the
default down the line if we decide to apply a different domain name to
the CI instance.