Refactor Jenkins user class
We need to include the user class on Jenkins agent machines, which
require the user and home directory created but do not need SSH keys or
gitconfig.

Since these two things are only needed on the master, move this into the
main class and just keep the user creation in it's own class to be
included elsewhere. Explicitly call in the jenkins_user and
jenkins_homedir parameters in the main class as these are used in
several places, so if we wanted to overwrite these on the master there
would be less changes.

Further, ensure the govuk_jenkins::pipeline class requires on the
govuk_jenkins::user class rather than a file resource.