Add PATH on Jenkinsfile
When trying to run builds on the new Jenkins agents using the swarm
client, we were receiving the following error:

```
jenkins.sh: bundle: not found
```

We need to explicitly set the path so it is able to run bundler. This
could be passed in a different way in the future but for now set it on
the Jenkinsfile.

This didn't occur on the Jenkins master because we believe environmental
variables are set by Jenkins itself.