Add option to pass in different env vars
If we need to move away from the default set environment variables which
are set elsewhere, we need to be able to break them out and specify them
by class as we are unable to differentiate by specific node class where
we set our secrets in hieradata.

This allows us to set a different PATH for the CI cluster.