Rework Jenkinsfile
The main goal of this commit is to put each of the build steps into the
Jenkinsfile rather than relying on a jenkins.sh shell script.

We are also removing the puppet lint publisher check. All this seems to
do is tell Github that "build succeeded, but lint failed!". We feel this
is over complicating what it offers, and people would be prone to just
check the console if it fails anyway. It also makes the Jenkinsfile
*much* easier to read.

Since each of the steps are now in their own stage, the pipeline view
splits up each stage so they can now be viewed individually in terms of
logs and timings: it will be much easier to see what's taking a long
time, and easier to parse log files.

Enable the pushing tags feature and remove jenkins.sh for migration
purposes.