Remove scheduled rake tasks env var
Initially I thought this would be a good workaround
to passing the necessary config to the application for
use with the 'whenever' gem. I've got doubts about this now
and attempts to test the signon application code needed
to process the value of this haven't worked out.
I intend to move it to a simpler cron config in another
pull request