Make CI agent Swarm property static to prevent dependency cycle
We found a dependency cycle when the fsroot property in the CI
agent Swarm class is configured to match the Jenkins home directory,
that is managed by other resources. If we want to enable the option
of having fsroot on a different directory we need to add some lines
to declare the directory only when it's different to the Jenkins home.
Until then we can leave fsroot static and expect this to be managed
by the Jenkins user.