Set fixed TCP port for Jenkins agent
This config setting allows you to set a fixed TCP port for connecting to
Jenkins agents rather than using a random port number. Fixing this
allows us to be more secure.

'0' is default and assigns a random number, as we do not need to set
this on Jenkins deployment machines.