Write MONGODB_NODES for manuals-publisher correctly
This value is expected to be slotted into a YAML file as an array.  The key
the value belongs to is nested and so the lines need to be indented correctly
or the resulting YAML file will be invalid.  Given some YAML as follows:

    production:
      hosts:
        - <%= ENV['MONGODB_NODES'] %>

If we don't add indentation to the lines after the first one what we end up
with is:

    production:
      hosts:
        - - first_node:27017
    - second_node:27017
    - and-so-on

This is not valid YAML and so the app will refuse to load the mongoid config
and ultimately refuse to run.  By tweaking the env var value to indent those
lines and correctly format the YAML as follows:

    production:
      hosts:
        - - first_node:27017
          - second_node:27017
          - and-so-on

And the app will boot again.

The bug in the formatting of this env var value was exposed by commit
336343d5198a62e76ab1da5515a6e40cbaf9fc0b where we introduced support
for multiline env vars.  Prior to this commit this value was ultimately treated
as a single line value.  The single line version of the value fits into the
YAML file correctly as the first line doesn't need indenting, so the problem
never arose.  Although it did mean that the app only ever communicated with
one of the expected mongo nodes.