Template config file for Jenkins 2
Jenkins 2 is super picky about its config file. Jenkins 2 will decide to
modify its config file when it starts so that it's formatted the way it
wants it to be. Puppet will then revert that change and cause Jenkins to
restart.

Here are some of the highlights of this ridiculousness:

- In Jenkins 2, things are sorted in some arbitrary way which is different
  to Jenkins 1. Don't try and sort things the old way. I could get on board
  with this except that it's less alphabetical than the sort order we were
  using in Jenkins 1.
- Jenkins 1 will add an exclamation mark before the XML for user IDs. Jenkins
  2 will remove that exclamation mark.
- If Jenkins 2 decides that it needs to rewrite your perfectly fine config file,
  it will write out a POSIX-incompatible file because it doesn't have trailing
  newline character.

I am sick of XML and I am sick of using Jenkins and I feel awful for littering
an XML template with conditionals but what are you going to do ¯\_(ツ)_/¯.