Remove reverse proxy for drug-analysis-prints
This relates to https://govuk.zendesk.com/agent/tickets/1453127

We're removing the reverse proxy configuration for `/drug-analysis-prints/` as MHRA have a new site which they want to manage mappings for.

The PDFs under `/home/groups/public/documents/sentineldocuments/` appear to all 404 now so we're removing this now.  I've requested that MHRA add mappings for these as there are 1436 of these without mappings.