Configure GitHub branch source plugin in CI to work with GitHub Enterprise
In the new CI Jenkins we want to manage GitHub Enterprise repositories
with the GitHub branch source plugin, same as we do with the GitHub
projects. This plugin provides better integration to update the status
of the build and we can configure all projects in the same way. The
GitHub Enterprise token credentials need to exist in Jenkins, and the
server needs to be added to the Global configuration.

When we use the GitHub branch source plugin for GitHub Enterprise, in addition
to adding the server certificate to the Java keystore, we also need to add it
to the machine certificate store. This is implemented with a new Puppet class
that manages the original code for the Java keystore and the new resources
for the OS certificate store.