Fix default content schema branch in Jenkins config
The content schema checkout step takes a git branch, with the default
'deployed-to-production'. In the old shell script, the default was
inserted using shell parameter expansion. In the new Groovy script,
the default should be inserted using a default method argument.

The implementation before this commit gave the error "Unimplemented
pathspec magic" because the ":-" characters were being passed to git.