Add a way to blacklist packages
We currently blacklist packages for unattended_upgrades, however when we
bring up a machine Puppet will install packages, and some may install
packages we do not want installed. This change makes it impossible to
install the specified packages whatsoever.

An example of this is systemd, which affects the way the system behaves
which we may not want to do yet. A newer version of postgresql installs
systemd despite it not being a dependency which caused some problems.