Add a way to blacklist packages
We currently blacklist packages for unattended_upgrades. However, when we
bring up a new machine, Puppet will install packages from fresh, and some packages
may install other packages we do not want installed. This change makes it impossible to
install the specified packages whatsoever.

An example of this is systemd, which affects the way the system behaves
which we may not want to do yet. A newer version of postgresql installs
systemd which caused some problems. We want to block this from
happening.