Add a default variable for ipaddress_eth0
A number of our puppet classes had their use of the `ipaddress` fact
changed to `ipaddress_eth0` when we introduced docker because facter was detecting `docker0` and using
the IP address that returned. In order for our tests to run under `strict_variables` we need to provide
a default for `ipaddress_eth0` too. I've picked 127.0.0.1 as it's a valid ipaddress and no
other significent reason.