Add ssh_slave defined type
This commit creates a defined type that should be used on a Jenkins
master that creates the node config to enable adding agents using the
ssh_slave plugin. This is a replacement of the Jenkins swarm method.

One prerequisite is to generate the "credentials_id" on the Jenkins UI
as this is directly mapped to the Jenkins instance we are unable to
programtically generate this.