Add relative back in, but do it from /
The `--relative` flag was removed because it caused incorrect pathnames.
However, this flag has a very handy function of creating the remote
directory structure. This is something which is not performed by
default.

We set the `--relative` flag back in to ensure files are copied, but set
the path to `/` so that they are copied to the correct location.