Fix signon rake tasks on Deploy Integration
We need to duplicate the signon deployable app for now to fix the
rake tasks on Deploy Integration (and Production). For now we are
adding a second job with a different name and same repository.

Once everything is tested with the new job, we can remove the original
signonotron2 one from the list of deployable apps.