Correctly docs.publishing.service.gov.uk to docs S3
In https://github.com/alphagov/govuk-puppet/pull/5269 I introduced a
nginx config to proxy docs.publishing.service.gov.uk to an S3 bucket.
The configuration was wrong, which caused nginx to crash loop on the
load balancers. Sad! The crash was fixed in
https://github.com/alphagov/govuk-puppet/pull/5279.

This time, I have employed a novel development technique called
"testing your stuff locally" to make sure that this config actually
works.

This commit fixes the following:

- By default, the nginx configs puts a HTTP basic auth password on the
proxy. We like the domain to be publicly accessible, so we disable this.
- To make S3 respond with the correct content, we set the `Hostname`
HTTP header to the name of the bucket. If this isn't set, Amazon will
try to find a bucket called `docs.publishing.service.gov.uk` and return
404.
- Add `docs` to the hostnames. This will make nginx respond on the new
`docs.x` domain.

I tested it by doing the following. First, start a load balancer VM on
locally with vagrant:

```
~/govuk/govuk-puppet $ vagrant up backend-lb-1.backend
```

Next, point the test domain to the running VM:

```
10.1.3.101 docs.dev.gov.uk
```

After this, visiting the docs will do the correct thing:

http://docs.dev.gov.uk/