Improve parameter name in Jenkins deployment helper
The integration deployment function in the pipeline helper takes an
application parameter - this needs to be the application ID configured
in puppet, not the repository. These are usually the same thing, but in
some applications they are different (e.g. business-support-finder vs
businesssupportfinder).

This change makes it clearer that the application ID is what's required.