Remove the stage() wrapper in the deployIntegration() library method
It seems like a convention is emerging that the Jenkinsfile in an
individual project is responsible for defining the pipeline stages.

Projects that used this library method are showing multiple 'deploy to
integration' stages, as this defined an additional stage internally.

Not a big deal if people don't want this change to go through, but it
will improve the pipeline view for a few projects.