Add way to safely restart Jenkins service
When we manage files on disk, we have to restart the Jenkins service if
we make a change. Doing a service refresh does not take into account if
any builds may be happening and will cause them to fail.

This commit adds a class which pulls the jenkins-cli.jar and issues a
safe restart. Use this class to notify rather than the service itself.