Add Jenkins helper function for validating whether to build a branch
Extracts a function from finder-frontend for identifying whether a
branch should be built, based on the branch name and a schema test flag.

Some branches, such as deployment branches, should not normally be built
by Jenkins because it's redundant and wasteful - that commit would have
already been tested on the master branch. But sometimes we do want to
build release branches, in order to test them against changes to the
content schema.