Delete unnecessary Jenkins helper
Delete the 'schemaTestParameters' function from the Jenkinsfile helper.

This function was intended to simplify Jenkinsfiles for projects which
are used as content schema tests, but it turned out not to be very
useful - it hides build parameters from the Jenkinsfile, and the
Jenkinsfile needs to know the parameters so that it can ensure their
default values are initialized.

The function is not being used by any applications, so it is safe to
delete.