Override Jenkins branches built for schema test repos
Ensure that 'deployed-to-production' is included in the Jenkins branch
builds. This is not normally required, but it is necessary for these
repos because the 'deployed-to-production' branch is used to test
changes pushed to the govuk-content-schemas repo.

This is just a sub-set of the repos which are built to test the content
schemas. Others will be configured in a similar way once their
Jenkinsfiles handle builds of deployed-to-production correctly.