Prevent Varnish from removing an example A/B testing cookie
Previously, all cookies were stripped from requests and responses, to
prevent applications (or their dependencies) adding cookies that users
had not been informed about, with the exception of the licence finder.

In order to A/B test changes to the navigation, we will need to add a
cookie to allow us to ensure a user always sees the same version of the
page.

This commit allows an example A/B testing cookie named 'abtest-example'
to pass through Varnish. This will be used to develop A/B testing on
Integration. Once that is done, it will be replaced with the name of the
actual A/B testing cookie, or a configured whitelist of cookies.