Do a hiera lookup instead of hardcoding
This fingerprint is hardcoded in the deployment repo as well as this
repo, and having it in two places made it less obvious that it had to be
changed elsewhere. Along with writing some documentation, do a hiera
lookup from the one source of truth so this is not missed in the future.