Dynamically lookup trade tariff hostname
It turns out that nginx looks up the hostname of a proxy_pass only
when it starts up, which means that if nginx hasn't restarted for
a while and Amazon changes the location of an ELB, we'll send
requests to the wrong place.

We've seen this before in ef64dda2105e7a055248d9dd9a7151315fe8acfe.