Enable new nav pages on integration
We will be releasing the new GOV.UK navigation pages soon. These pages
live in `collections` and in the meantime we would like to have them
available on integration. In order to do so, we have hidden the nav
pages behind a feature flag.

This commit adds an environment variable to collections on integration
so that we can show those new pages on only this environment.