Enable tagging documents to new taxonomy on integration
We will be releasing a new page on Whitehall to allow Education documents to be tagged to the new taxonomy.
In the meantime we would like to test them under integration environment.
To allow us to do that we have hidden the page behind a feature flag, so it only shows while on integration.

This commit adds an environment variable to whitehall on integration so that we can show the page only under this environment.