Purge mirrorer rabbitmq purge daily
We run govuk_crawler_worker to test the functionality, and it adds
messages to rabbitmq but these are never processed because we do not
have mirrors in Integration.

This adds a function that purges the queue everyday so avoid get alerts
related to rabbitmq memory because the queue has too many messages.