Override Jenkins branches built for schema test repos
By default, all of the deployment branches are excluded from Jenkins
build, because it normally doesn't make sense to build them. However, we
sometimes do want to build deployed-to-production to test a project
against changes to the content schemas.

This overrides the excluded branches for email-alert-service and
email-alert-frontend so that most of the release and deployment branches
are still excluded, but deployed-to-production can be built for schema
tests.