Adjust high watermark threshold
The threshold appears to be set too low for the rabbitmq runtime.  This causes the Nagios checks to raise alerts in Icinga (false positives) for integration.

What the threshold should be set to, is debatable. We are currently using the default value of 40% of resident memory for the mirrorer vms.
See: https://www.rabbitmq.com/memory.html

Alternatively, we could increase the memory of the VM to same as in production.

Rabbitmq consistently ocupies currently set threshold for the high watermark