Add env var for Appsignal trial
We're evaluating Appsignal as an alternative to Errbit. This adds a env
var that will be accessible to all applications. This works because
Appsignal uses one API key per account (as opposed to app and
environment for Errbit).

We'll only configure this in integration for our trial. If we decide to
go with another option we'll remove this.

https://trello.com/c/J3kDeZD7