Configure Jenkins not to merge `master` when building merged branches
Many Jenkins builds call the `mergeMasterBranch()` helper near the start
of the build. This alerts us (and GitHub) to branches which cannot be
cleanly merged with master. It uses the `--no-commit` flag to just check
whether a merge would be valid, without actually performing the merge.

This works fine in the mainstream use-case, which is to test dev
branches which have yet to be merged. However, it doesn't do the right
thing for content schema tests, where we are running builds on the
`deployed-to-production` branch. Unlike a dev branch, this branch is
_behind_ `master`. So the merge being tested in this case is a
fast-forward merge. Git doesn't apply the `--no-commit` behaviour for
fast-forward merges, so it actually performs the merge! This means that
we end up building the `master` branch rather than
`deployed-to-production`.

The consequence of this is that the content schema tests were not strict
enough: they were unintentionally running against the `master` version
of the downstream repos, not `deployed-to-production`, which is the
branch that we really care about breaking.

This fixes the problem by preventing the merge with `master` if the
current commit is in the history for `master`. This means that building
`deployed-to-production` for a schema test or manually triggering a
build of an old, merged dev branch will build exactly that branch rather
than (confusingly) building `master` instead.