Remove smokey for performance platform
This job has  been failing for almost 18 months, 4 times a day. It has
never run successfully:

https://deploy.publishing.service.gov.uk/job/Smokey_Performance_Platform
/1/

The long failure without somebody caring or noticing is evidence enough
for me that the tests wouldn't add value. This PR is easily reverted if
it turns out we need to support the tests again.