Use etcd on docker-management-1 for reboots
We created a container for etcd in
8e9f59a34e7c3f698f8acc752edd8fadd7f9d8d2. The main advantage is that it
enables us to easily upgrade to the latest version of etcd, and also
remove a bunch of surrounding etcd code used to manage the current etcd
machine.

This adds a service alias that will become "etcd.cluster", and update
the cluster endpoint to use for unattended reboots in Integration only.

Integration is a little tricky to test as the machines are shut down every night,
but we could potentially change the times machines are rebooted.