Ensure content-performance-manager absent in prod
The content-performance-manager is a new app, and is not ready to be
deployed in production.

When the original puppet code was added for the application, the feature
flag was missing, so all of the alerts and checks were set up even though
the application does not exist.

I tried rectifying this by adding the feature flag, but this didn’t
remove the framework that was already put in place.

Setting `ensure` to absent is one of the first steps in
[retiring an application](https://github.gds/pages/gds/opsmanual/infrastructure/howto/retiring-an-application.html). I’m hoping that this will remove all of
the alerts for content-performance-manager

I’ve removed the feature flag because the value for `ensure` can’t be set
while it exists.