Make sure env vars are only created if app is present
Puppet won’t allow you to create environment variables for an app that
does not have any records (i.e. is ‘absent’)

The last run of puppet threw errors for content-performance-manager,
because the manifest was trying to create environment variables for an
application framework that doesn’t exist:

```
[backend-3.backend] out: Error: Could not set 'present' on ensure: No such file or directory - /etc/govuk/content-performance-manager/env.d/GOOGLE_PRIVATE_KEY20170130-4949-gakref.lock at 36:/usr/share/puppet/production/current/modules/govuk/manifests/app/envvar.pp
[backend-3.backend] out: Error: Could not set 'present' on ensure: No such file or directory - /etc/govuk/content-performance-manager/env.d/GOOGLE_PRIVATE_KEY20170130-4949-gakref.lock at 36:/usr/share/puppet/production/current/modules/govuk/manifests/app/envvar.pp
```

Wrapped the creation of the environment variables in a conditional to fix
this.