Add post data sync tasks for content-performance-manger
content-performance-manager is not deployed in production. This means
that every time the data sync jobs to staging and integration run, their
databases are copied over with a blank database that does not have any
tables.

These post sync jobs, re-deploy the app and run migrations and 
populate the databases with minimal data.

When the application is finally deployed to production, these jobs will
be removed.