Set the `package_ensure` parameter on `mysql::client`
Our vendored version of the mysql module has a bug that is
addressed in
https://github.com/puppetlabs/puppetlabs-mysql/commit/18df1d08e5c326b1ad91b300864b30e6087d9ec4
it attempts to use a value that `params` does not define.

We have two main options for working around this (as we don't have time to
do a module upgrade at this point).

1: fork the upstream and apply the one change to fix the bug
2: explicitly set the param using hiera

Option 1 adds another module to our collection of forks and moves us
slightly further away from upstream. We'd rather not incur this cost.

Option 2 Adds hiera config that we shouldn't technically need but allows
us to retain our current code.

This PR takes options 2 and ensures the setting is present in both tests
and our running config